+++ app/controllers/users_controller.rb
@@ -3,41 +3,13 @@
 # Allows admins to manage user accounts.
 class UsersController < ApplicationController
   before_action :authenticate_user!
-  before_action :check_system_admin, except: [:settings, :update_settings, :change_password]
+  before_action :check_system_admin
   before_action :set_user, only: [:show, :edit, :update, :destroy]
   before_action :redirect_without_user, only: [:show, :edit, :update, :destroy]
 
-  # # GET /settings
-  # def settings
-  # end
-
-  # PATCH /settings
-  def update_settings
-    if current_user.update(user_params)
-      redirect_to settings_path, notice: 'Settings successfully updated.'
-    else
-      render :settings
-    end
-  end
-
-  # PATCH /change_password
-  def change_password
-    if current_user.valid_password?(params[:user][:current_password])
-      if current_user.reset_password(params[:user][:password], params[:user][:password_confirmation])
-        bypass_sign_in current_user
-        redirect_to settings_path, notice: 'Your password has been changed.'
-      else
-        render :settings
-      end
-    else
-      current_user.errors.add(:current_password, 'is invalid')
-      render :settings
-    end
-  end
-
   # GET /users
   def index
-    @order = scrub_order(User, params[:order], 'users.current_sign_in_at desc')
+    @order = scrub_order(User, params[:order], "users.current_sign_in_at desc")
     @users = User.current.search(params[:search]).order(@order).page(params[:page]).per(40)
   end
 
@@ -48,7 +20,7 @@ class UsersController < ApplicationController
   # PATCH /users/1
   def update
     if @user.update(user_params)
-      redirect_to @user, notice: 'User was successfully updated.'
+      redirect_to @user, notice: "User was successfully updated."
     else
       render :edit
     end
@@ -59,7 +31,7 @@ class UsersController < ApplicationController
   def destroy
     @user.destroy
     respond_to do |format|
-      format.html { redirect_to users_path, notice: 'User was successfully deleted.' }
+      format.html { redirect_to users_path, notice: "User was successfully deleted." }
       format.js
     end
   end
@@ -75,17 +47,11 @@ class UsersController < ApplicationController
   end
 
   def user_params
-    params[:user] ||= {}
-    if current_user.system_admin?
-      params.require(:user).permit(
-        :first_name, :last_name, :email, :username, :research_summary,
-        :degree, :aug_member, :core_member, :system_admin, :community_manager,
-        :banned, :emails_enabled, :contributor
-      )
-    else
-      params.require(:user).permit(
-        :first_name, :last_name, :email, :username, :research_summary, :emails_enabled
-      )
-    end
+    params.require(:user).permit(
+      :first_name, :last_name, :email, :username, :research_summary, :degree,
+      :aug_member, :core_member, :system_admin, :community_manager, :banned,
+      :emails_enabled, :contributor, :profile_bio, :profile_url,
+      :profile_location
+    )
   end
 end