+++ test/controllers/internal_controller_test.rb
@@ -3,32 +3,20 @@
 require "test_helper"
 
 # Tests for internal user pages.
-class InternalControllerTest < ActionController::TestCase
-  test "should get dashboard as regular user" do
-    login(users(:regular))
-    get :dashboard
-    assert_response :success
+class InternalControllerTest < ActionDispatch::IntegrationTest
+  setup do
+    @regular = users(:regular)
   end
 
-  # test "should get settings" do
-  #   get :settings
-  #   assert_response :success
-  # end
-
-  # test "should get tools" do
-  #   get :tools
-  #   assert_response :success
-  # end
-
-  test "should get profile" do
-    login(users(:regular))
-    get :profile
+  test "should get dashboard" do
+    login(@regular)
+    get dashboard_url
     assert_response :success
   end
 
   test "should get token" do
-    login(users(:regular))
-    get :token
+    login(@regular)
+    get token_url
     assert_response :success
   end
 end